Add activate_switchboard() in Building class
activate_switchboard() will control the elevator giving it
a cue of floor to travel to. This required a cue to be added
to Elevator class, and changes in the behaviour of run_elevator().
Also, removed all graphic cues to debugging. (Can still be used
in old version branch).
Also, removed all statistical logs and figures. (Can still be used
in old version branch).